moved partner logos component to about page
